Add fastfetch to servers packages
While neofetch was mostly a "fun" tool, fastfetch reached a level of customization and additional system information to the point where I feel like it's actually useful to also have it on servers.

Indeed, in addition of the usual information like the running kernel, system uptime and so on... fastfetch can also show the CPU temp, the locale, the ip addresse(s), disk space of mount points, etc... It's a nice way to wrap a lot of useful information into a single command.

I added fastfetch in Debian/Ubuntu procedure despite the fact it is not available yet in the repos but I guess [it should land there soon-ish](https://salsa.debian.org/debian/fastfetch).